+++ CHANGES.rst
@@ -34,6 +34,14 @@ Release history
 - Added ``LinearSystem``, a ``Process`` to implement arbitrary linear systems.
   (`#1650`_)
 
+**Changed**
+
+- ``LinearFilter`` now accepts system descriptions in state-space (``(A, B, C, D)``) or
+  zero-pole-gain (``(z, p, k)``) form, in addition to transfer function (``(num, den)``)
+  form. If passing systems in transfer function form, they should be passed as a
+  ``(numerator, denominator)`` tuple; the separate ``den`` argument is deprecated.
+  (`#1650`_)
+
 **Fixed**
 
 - Operator graph step order will now be deterministic. (`#1654`_)